python/modular-version/v0.1/requirements.txt
flask==3.0.0
supabase==2.0.3
python-dotenv==1.0.0
requests==2.31.0
werkzeug==3.0.1
flask-cors==4.0.0
flask-socketio==5.3.6
python-engineio==4.8.2
python-socketio==5.11.1
